lazy commit

+++ sorting_algorithms/sort.h
@@ -5,7 +5,7 @@
 #define SORT_H
 #include <stdio.h>
 
-int bubble_sort(int argc, unsigned long *args);
+int bubble_sort(int argc, int *args);
 int merge_sort(int argc, unsigned long args[]);
 int insertion_sort(int argc, unsigned long args[]);
 int tim_sort(int argc, unsigned long args[]);
